requirements.txt
bitcoin==1.1.42
eth-abi==1.3.0
eth-utils==1.4.1
pysha3==1.0.2
requests==2.20.1
rlp==0.6.0

